+++ src/main/java/ghidra/emotionengine/symboltable/StdumpAST.java
@@ -52,6 +52,7 @@ public class StdumpAST {
 		HashMap<String, Integer> typeNameToDeduplicatedTypeIndex = new HashMap<>();
 		String conflictResolutionPostfix;
 		boolean hadBadTypeLookup = false;
+		HashMap<String, StructureDataType> forwardDeclaredTypes = new HashMap<>();
 		
 		// Ghidra objects.
 		TaskMonitor monitor;
@@ -459,12 +460,7 @@ public class StdumpAST {
 			}
 			Integer index = lookupTypeIndex(importer);
 			if(index == null) {
-				if(!importer.hadBadTypeLookup) {
-					importer.log.appendMsg("STABS", "Type lookup failures are normal in cases where a type is forward declared in a translation unit with symbols, but is not defined in one.");
-					importer.hadBadTypeLookup = true;
-				}
-				importer.log.appendMsg("STABS", "Type lookup failed: " + typeName);
-				return Undefined1DataType.dataType;
+				return createForwardDeclaredType(importer);
 			}
 			DataType type = importer.types.get(index);
 			if(type == null) {
@@ -475,6 +471,21 @@ public class StdumpAST {
 				}
 				type = node.createType(importer);
 				importer.types.set(index, type);
+			}
+			return type;
+		}
+		
+		public DataType createForwardDeclaredType(ImporterState importer) {
+			if(!importer.hadBadTypeLookup) {
+				importer.log.appendMsg("STABS", "Type lookup failures are normal in cases where a type is forward declared in a translation unit with symbols, but is not defined in one.");
+				importer.hadBadTypeLookup = true;
+			}
+			importer.log.appendMsg("STABS", "Type lookup failed: " + typeName);
+			StructureDataType type = importer.forwardDeclaredTypes.get(typeName);
+			if(type == null) {
+				type = new StructureDataType(typeName, 1, importer.programTypeManager);
+				type.setDescription("Probably forward declared, but not defined, in a translation unit with symbols.");
+				importer.forwardDeclaredTypes.put(typeName, type);
 			}
 			return type;
 		}