+++ src/main/java/ghidra/emotionengine/symboltable/StdumpAST.java
@@ -53,6 +53,7 @@ public class StdumpAST {
 		String conflictResolutionPostfix;
 		boolean hadBadTypeLookup = false;
 		HashMap<String, StructureDataType> forwardDeclaredTypes = new HashMap<>();
+		ArrayList<String> prefixStack = new ArrayList<>(); // Used to name nested structs.
 		
 		// Ghidra objects.
 		TaskMonitor monitor;
@@ -89,7 +90,6 @@ public class StdumpAST {
 		boolean conflict = false;
 		int stabsTypeNumber = -1;
 		
-		String prefix = ""; // Used for nested structs.
 		boolean isInCreateTypeCall = false; // Prevent infinite recursion.
 		
 		public DataType createType(ImporterState importer) {
@@ -116,11 +116,21 @@ public class StdumpAST {
 			}
 		}
 		
-		String generateName(ImporterState importer) {
-			if(name == null || name.isEmpty()) {
-				return prefix + "__unnamed_" + Integer.toString(absoluteOffsetBytes) + importer.conflictResolutionPostfix;
+		String generateName(ImporterState importer, boolean addPrefix) {
+			String prefixString = "";
+			if(addPrefix) {
+				for(String prefixPart : importer.prefixStack) {
+					if(prefixPart != null) {
+						prefixString += prefixPart + "__";
+					} else {
+						prefixString += "unnamed__";
+					}
+				}
 			}
-			return prefix + name + importer.conflictResolutionPostfix;
+			if(name == null || name.isEmpty()) {
+				return prefixString + "unnamed_" + Integer.toString(absoluteOffsetBytes) + importer.conflictResolutionPostfix;
+			}
+			return prefixString + name + importer.conflictResolutionPostfix;
 		}
 	}
 	
@@ -233,7 +243,7 @@ public class StdumpAST {
 		ArrayList<EnumConstant> constants = new ArrayList<EnumConstant>();
 		
 		public DataType createTypeImpl(ImporterState importer) {
-			EnumDataType type = new EnumDataType(generateName(importer), 4);
+			EnumDataType type = new EnumDataType(generateName(importer, true), 4);
 			for(EnumConstant constant : constants) {
 				type.add(constant.name, constant.value);
 			}
@@ -249,12 +259,12 @@ public class StdumpAST {
 		
 		public DataType createTypeImpl(ImporterState importer) {
 			DataType result = createEmpty(importer);
-			fill(result, 0, this, importer);
+			fill(result, 0, this, this, importer);
 			return result;
 		}
 		
 		public DataType createEmpty(ImporterState importer) {
-			String typeName = generateName(importer);
+			String typeName = generateName(importer, true);
 			int sizeBytes = sizeBits / 8;
 			DataType type;
 			if(isStruct) {
@@ -265,64 +275,63 @@ public class StdumpAST {
 			return type;
 		}
 		
-		public void fill(DataType dest, int baseOffset, InlineStructOrUnion topLevelNode, ImporterState importer) {
-			if(isStruct) {
+		public static void fill(DataType dest, int baseOffset, InlineStructOrUnion baseClassNode, InlineStructOrUnion subClassNode, ImporterState importer) {
+			if(baseClassNode.isStruct) {
 				Structure type = (Structure) dest;
-				for(int i = 0; i < baseClasses.size(); i++) {
+				for(int i = 0; i < baseClassNode.baseClasses.size(); i++) {
 					if(importer.embedBaseClasses) {
-						InlineStructOrUnion baseClass = lookupBaseClass(i, importer);
+						Node baseClassTypeName = baseClassNode.baseClasses.get(i);
+						InlineStructOrUnion baseClass = baseClassNode.lookupBaseClass(i, importer);
 						if(baseClass == null) {
 							continue;
 						}
-						baseClass.fill(dest, baseClass.absoluteOffsetBytes, topLevelNode, importer);
+						importer.prefixStack.add(baseClassNode.name);
+						fill(dest, baseOffset + baseClassTypeName.absoluteOffsetBytes, baseClass, subClassNode, importer);
+						importer.prefixStack.remove(importer.prefixStack.size() - 1);
 					} else {
-						Node baseClass = baseClasses.get(i);
+						Node baseClass = baseClassNode.baseClasses.get(i);
 						DataType baseType = replaceVoidWithUndefined1(baseClass.createType(importer));
 						String baseClassName = "base_class_" + Integer.toString(baseClass.absoluteOffsetBytes);
 						if(baseClass instanceof TypeName) {
 							baseClassName += "_" + ((TypeName) baseClass).typeName;
 						}
-						addField(type, baseType, baseClass, baseClass.absoluteOffsetBytes, baseClassName, importer);
+						addField(type, baseType, baseClass, baseOffset + baseClass.absoluteOffsetBytes, baseClassName, false, baseClassNode, subClassNode, importer);
 					}
 				}
-				for(Node node : fields) {
+				for(Node node : baseClassNode.fields) {
 					if(node.storageClass != StorageClass.STATIC) {
 						// Currently we don't try to import bit fields.
 						DataType field = null;
 						if(node.name != null && node.name.equals("__vtable")) {
-							field = new PointerDataType(topLevelNode.createVtable(importer));
+							field = new PointerDataType(subClassNode.createVtable(importer));
 						} else {
-							if(prefix != null) {
-								node.prefix += prefix;
-							}
-							if(name != null) {
-								node.prefix += name + "__";
-							}
+							importer.prefixStack.add(baseClassNode.name);
 							field = replaceVoidWithUndefined1(node.createType(importer));
+							importer.prefixStack.remove(importer.prefixStack.size() - 1);
 						}
-						addField(type, field, node, node.relativeOffsetBytes, node.name, importer);
+						boolean isInherited = baseClassNode != subClassNode;
+						addField(type, field, node, baseOffset + node.relativeOffsetBytes, node.name, isInherited, baseClassNode, subClassNode, importer);
 					}
 				}
 			} else {
+				if(baseClassNode != subClassNode) {
+					importer.log.appendMsg("baseClassNode != subClassNode for union");
+				}
 				Union type = (Union) dest;
-				for(Node node : fields) {
+				for(Node node : baseClassNode.fields) {
 					if(node.storageClass != StorageClass.STATIC) {
-						if(prefix != null) {
-							node.prefix += prefix;
-						}
-						if(name != null) {
-							node.prefix += name + "__";
-						}
+						importer.prefixStack.add(baseClassNode.name);
 						DataType field = replaceVoidWithUndefined1(node.createType(importer));
+						importer.prefixStack.remove(importer.prefixStack.size() - 1);
 						type.add(field, field.getLength(), node.name, "");
 					}
 				}
 			}
 		}
 		
-		public void addField(Structure structure, DataType field, Node node, int offset, String fieldName, ImporterState importer) {
+		public static void addField(Structure structure, DataType field, Node node, int offset, String fieldName, boolean isInherited, InlineStructOrUnion baseClassNode, InlineStructOrUnion subClassNode, ImporterState importer) {
 			boolean isBitfield = node instanceof BitField;
-			boolean isBeyondEnd = offset + field.getLength() > sizeBits / 8;
+			boolean isBeyondEnd = offset + field.getLength() > subClassNode.sizeBits / 8;
 			boolean isZeroLengthStruct = false;
 			if(field instanceof Structure) {
 				Structure structField = (Structure) field;
@@ -331,8 +340,12 @@ public class StdumpAST {
 				}
 			}
 			if(!isBitfield && !isBeyondEnd && !isZeroLengthStruct) {
+				String comment = "";
+				if(isInherited) {
+					comment = "Inherited from " + baseClassNode.name;
+				}
 				try {
-					structure.replaceAtOffset(offset, field, field.getLength(), fieldName, "");
+					structure.replaceAtOffset(offset, field, field.getLength(), fieldName, comment);
 				} catch(IllegalArgumentException e) {
 					importer.log.appendException(e);
 				}
@@ -340,8 +353,9 @@ public class StdumpAST {
 		}
 		
 		public DataType createVtable(ImporterState importer) {
+			String vtableName = generateName(importer, false) + "__vtable";
 			int vtableSize = calculateVtableSize(importer);
-			StructureDataType vtable = new StructureDataType(generateName(importer) + "__vtable", vtableSize, importer.programTypeManager);
+			StructureDataType vtable = new StructureDataType(vtableName, vtableSize, importer.programTypeManager);
 			fillVtable(vtable, importer);
 			return vtable;
 		}